Uniformed initing of GraphService for POST and GET
Now optical system is given for GrapshService in both cases
Optical system is fetched for POST in appropriate strategies, for GET directly in controller
Optical system fetching is implemented with opt_sys_id, which is stored in cookies. If not stored, a new opt sys is provided.
The usage of uniformed context and uniformed service initing made try-except checking be required in a case of optical system has not been provided yet